validateModule
