+++ regExp.js
@@ -1,7 +1,5 @@
 const fs = require('fs');
-const path = require('path');
-const fetch = require('node-fetch');
-const fileReadme = './README.md';
+const path = require('path')
 
 const pathText = path.join(__dirname, './readme.md');
 const readMe = fs.readFileSync(pathText, 'utf-8');
@@ -10,19 +8,9 @@ const readMe = fs.readFileSync(pathText, 'utf-8');
 const regex = /\[(.+)\]\((https?:.+)\)/ig;
 const linksArray = Array.from(readMe.match(regex))
 
-const newArr =linksArray.map(texto => {
-    const splitText = texto.split('(');
-    const linkText = splitText[0].replace('[', '').replace(']', '');
-    const replaceText = splitText.pop().replace(')', '')
-return {href:replaceText, text:linkText, file: fileReadme}
 
-    /*fetch(replaceText).then(response => {
-        console.log(response.status)
-    });*/
+module.exports = { linksArray }
 
-});
-//console.log(linksArray);
-console.log(newArr);
